Move components to src root
