+++ lib/src/main/java/com/zetcode/components/PasswordEx.java
@@ -1,4 +1,4 @@
-package com.zetcode;
+package com.zetcode.components;
 
 import javax.swing.AbstractAction;
 import javax.swing.GroupLayout;
