+++ lib/src/main/java/com/zetcode/components/cpu.png
