+++ lib/src/main/java/com/zetcode/components/home.png
