+++ lib/src/main/java/com/zetcode/components/laptop.png
